refactor: enhance drawer component structure and functionality
