+++ docs/src/app/layouts/BaseLayout.astro
@@ -88,12 +88,12 @@ const seoMeta = buildSeoMeta({
       currentPath={currentPath}
     />
 
-    <div class:list={["block w-full min-h-[calc(100vh-var(--header-height,77px))]", toc !== "off" && "wide:flex"]} id="page-layout">
+    <div class:list={["block w-full min-h-[calc(100dvh-var(--header-height,77px))]", toc !== "off" && "wide:flex"]} id="page-layout">
       {toc !== "off" && <TocSidebar tocMode={toc} headings={headings} />}
 
       <main
         class:list={[
-          "main-content min-h-[calc(100vh-200px)] py-lg mobile:py-md print:py-0",
+          "main-content min-h-[calc(100dvh-200px)] py-lg mobile:py-md print:py-0",
           toc !== "off" && "wide:min-w-0 wide:flex-[1_1_0%]",
           pageSlug === "home" && "pt-md",
         ]}
