+++ docs/src/shared/lib/drawer/index.ts
@@ -1,0 +1,2 @@
+export { createDrawer } from "./model/createDrawer";
+export type { DrawerConfig, DrawerHandle } from "./model/types";
