+++ docs/src/shared/ui/tailwind.ts
@@ -799,24 +799,23 @@ export const TOC_TOGGLE_CLASS = [
   "items-center",
   "justify-center",
   "rounded-full",
-  "border",
-  "border-border",
-  "bg-surface",
-  "text-text",
+  "bg-primary",
+  "text-text-inverse",
   "shadow-sm",
-  "transition-all",
+  "pointer-events-auto",
+  "transition-[background,transform,opacity]",
   "duration-200",
-  "ease-out",
-  "hover:-translate-y-0.5",
-  "hover:border-border",
-  "hover:text-primary",
+  "ease-[var(--transition-fn)]",
+  "hover:scale-[1.06]",
   "hover:shadow-md",
-  "active:translate-y-0",
+  "hover:bg-primary-hover",
+  "[html.dark-mode_&]:bg-[linear-gradient(135deg,var(--color-header-bg),var(--color-header-bg-end))]",
+  "[html.dark-mode_&]:hover:bg-[linear-gradient(135deg,var(--color-header-bg-end),var(--color-header-bg))]",
   "active:scale-95",
   "data-[nav-hidden=true]:pointer-events-none",
   "data-[nav-hidden=true]:opacity-0",
   "wide:hidden",
-  FOCUS_RING_CLASS,
+  INVERSE_FOCUS_RING_CLASS,
 ].join(" ");
 
 export const TOC_PANEL_CLASS = [
@@ -828,20 +827,25 @@ export const TOC_PANEL_CLASS = [
   "h-[calc(100dvh-var(--header-height))]",
   "w-[260px]",
   "translate-x-[-100%]",
+  "opacity-0",
   "flex-col",
   "border-r",
   "border-border-light",
   "bg-surface",
-  "transition-[transform,box-shadow]",
+  "pointer-events-auto",
+  "will-change-transform",
+  "transition-[transform,opacity,box-shadow]",
   "duration-toc",
   "ease-[cubic-bezier(0.4,0,0.2,1)]",
   "[.toc-sidebar.is-open_&]:translate-x-0",
+  "[.toc-sidebar.is-open_&]:opacity-100",
   "[.toc-sidebar.is-open_&]:shadow-lg",
   "wide:sticky",
   "wide:top-[calc(var(--header-height)+1rem)]",
   "wide:h-[calc(100vh-var(--header-height)-1rem)]",
   "wide:w-full",
   "wide:translate-x-0",
+  "wide:opacity-100",
   "wide:rounded-t-lg",
   "wide:shadow-none",
 ].join(" ");
