+++ docs/src/widgets/home/ui/Hero.astro
@@ -84,7 +84,7 @@ const heroNoteLinkClass = [
       {hero.subtitle}
     </p>
 
-    <div class="flex flex-wrap justify-center gap-md" aria-labelledby="hero-heading">
+    <div class="flex flex-wrap justify-center gap-md mobile:w-full mobile:flex-col mobile:items-stretch" aria-labelledby="hero-heading">
       <a href={withBase(hero.primary_cta.url)} class={BUTTON_PRIMARY_CLASS}>{hero.primary_cta.label}</a>
       <a href={withBase(hero.secondary_cta.url)} class={BUTTON_SECONDARY_CLASS}>{hero.secondary_cta.label}</a>
       <a href={withBase(hero.community_cta.url)} class={BUTTON_OUTLINE_CLASS} target="_blank" rel="noopener noreferrer">
