+++ docs/src/widgets/site-shell/ui/Header.astro
@@ -125,7 +125,7 @@ const themeButtonClass = [
         </div>
 
         <nav
-          class="main-nav order-3 w-full overflow-hidden max-h-0 opacity-0 transition-[max-height,opacity] duration-nav ease-in-out mobile:flex-[1_0_100%] mobile:[&.is-open]:max-h-[540px] mobile:[&.is-open]:opacity-100 mobile:[&.is-closing]:pointer-events-none tablet:order-none tablet:flex-[1_1_auto] tablet:min-w-0 tablet:overflow-visible tablet:max-h-none tablet:opacity-100"
+          class="main-nav order-3 w-full overflow-hidden max-h-0 opacity-0 transition-[max-height,opacity,transform] duration-nav ease-in-out mobile:flex-[1_0_100%] mobile:-translate-y-1 mobile:[&.is-open]:max-h-[540px] mobile:[&.is-open]:opacity-100 mobile:[&.is-open]:translate-y-0 mobile:[&.is-closing]:pointer-events-none tablet:order-none tablet:flex-[1_1_auto] tablet:min-w-0 tablet:overflow-visible tablet:max-h-none tablet:opacity-100"
           id="main-nav"
           aria-label="Main navigation"
         >
